Breakdown stylesheet in several types toward plugin handlers
