+++ lib/files/file.js
@@ -2,7 +2,6 @@
 
 const fs              = require('fs-extra')
 const path            = require('upath')
-const pluralize       = require('inflection').pluralize
 // TODO: Evaluate removal of BaseFile class require by uncommenting code
 // ...depending on the "contexter module" API strategy
 // For example, here File extends BaseFile located inside module
@@ -74,10 +73,11 @@ module.exports = class File extends BaseFile {
 
     Object.assign(this.path, path.parse(this.path.relative))
 
+    // TODO: Move targetExt logic to each file type handler
     var targetExt = this.path.ext.toLowerCase()
     if (this.type === 'page') targetExt = '.html'
     if (this.type === 'script') targetExt = '.js'
-    if (this.type === 'stylesheet') targetExt = '.css'
+    if (this.type.indexOf('stylesheet') > -1) targetExt = '.css'
 
     var targetFull = this.path.full
       .replace(sourceDir, targetDir)
@@ -111,25 +111,16 @@ module.exports = class File extends BaseFile {
   write(context, done){
     fs.mkdirsSync(path.dirname(this.path.target.full))
 
-    switch(this.type) {
-      case 'page':
-      case 'script':
-      case 'stylesheet':
-        return this.render(context, (err, output) => {
-          if (err) throw err
-          fs.writeFileSync(this.path.target.full, output)
-          return done(null)
-        })
-        break
-      case 'layout':
-        // no-op
-        break
-      case 'datafile':
-      case 'image':
-      case 'unknown':
-      default:
-        fs.copySync(this.path.full, this.path.target.full)
-        break
+    if (this.isSqueezable) {
+      return this.render(context, (err, output) => {
+        if (err) throw err
+        fs.writeFileSync(this.path.target.full, output)
+        return done(null)
+      })
+    }
+
+    if (!this.isRemovable) {
+      fs.copySync(this.path.full, this.path.target.full)
     }
 
     return done(null)
@@ -139,7 +130,20 @@ module.exports = class File extends BaseFile {
     return this.path.relative
   }
 */
+// TODO: Move `isSqueezable()` and `isRemovable()` logic to each file type handler
   get isSqueezable() {
-    return ['page', 'script', 'stylesheet'].indexOf(this.type) > -1
+    return [
+      'page',
+      'script',
+      'stylesheetcss',
+      'stylesheetless',
+      'stylesheetsass',
+      'stylesheetscss',
+      'stylesheetstyl'
+    ].indexOf(this.type) > -1
+  }
+
+  get isRemovable() {
+    return ['layout'].indexOf(this.type) > -1
   }
 }
