+++ lib/module/context.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const findKey         = require('lodash').findKey
-// const detectSeries    = require('async').detectSeries
 const pluralize       = require('inflection').pluralize
 const extensions      = require('./extensions')
 const typer           = require('./typer')
@@ -33,7 +32,7 @@ module.exports = class Context {
 
     this.files =  [] // start with files array empty
     Object.keys(fileTypes).forEach(type => {
-      const t = pluralize(type)
+      const t = pluralize(type.toLowerCase())
       this[t] = [] // add keys and initialize to an empty array
     })
   }
@@ -41,6 +40,7 @@ module.exports = class Context {
   newFile (filename, sourceDir, targetDir) {
     var type
 
+    // `typer()` function available? if not, use its own handler `check()`
     if (opts && opts.typer) {
       type = opts.typer(filename)
     } else {
